had to wrap the file read in a filesystem existence check
If the codebase used an unqualified import from the core libraries, or from an Elm package library, then the code was going out to the user's file system looking for it and blowing up when it couldn't find the file. This would cause all the workers to die, and Elm Lens wasn't able to make the markup happen.